backend/requirements.txt
aiofiles==25.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
doclayout_yolo==0.0.4
fastapi==0.132.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
#opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
psutil==7.2.2
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
PyMuPDF==1.27.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scipy==1.17.1
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==82.0.0
shellingham==1.5.4
simsimd==6.5.13
six==1.17.0
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
